+++ engine/src/Updaters/Color/index.ts
@@ -1,6 +1,6 @@
 import type { Main } from "../../main";
 import { ColorUpdater } from "./ColorUpdater";
 
-export function loadColorUpdater(tsParticles: Main): void {
-    tsParticles.addParticleUpdater("color", (container) => new ColorUpdater(container));
+export async function loadColorUpdater(tsParticles: Main): Promise<void> {
+    await tsParticles.addParticleUpdater("color", (container) => new ColorUpdater(container));
 }