+++ engine/src/slim.ts
@@ -23,31 +23,31 @@ import { loadTextShape } from "./Shapes/Text";
 import { loadParticlesLinksInteraction } from "./Interactions/Particles/Links";
 import { loadSizeUpdater } from "./Updaters/Size";
 
-export function loadSlim(tsParticles: Main): void {
-    loadExternalAttractInteraction(tsParticles);
-    loadExternalBounceInteraction(tsParticles);
-    loadExternalBubbleInteraction(tsParticles);
-    loadExternalConnectInteraction(tsParticles);
-    loadExternalGrabInteraction(tsParticles);
-    loadExternalRepulseInteraction(tsParticles);
+export async function loadSlim(tsParticles: Main): Promise<void> {
+    await loadExternalAttractInteraction(tsParticles);
+    await loadExternalBounceInteraction(tsParticles);
+    await loadExternalBubbleInteraction(tsParticles);
+    await loadExternalConnectInteraction(tsParticles);
+    await loadExternalGrabInteraction(tsParticles);
+    await loadExternalRepulseInteraction(tsParticles);
 
-    loadParticlesAttractInteraction(tsParticles);
-    loadParticlesCollisionsInteraction(tsParticles);
-    loadParticlesLinksInteraction(tsParticles);
+    await loadParticlesAttractInteraction(tsParticles);
+    await loadParticlesCollisionsInteraction(tsParticles);
+    await loadParticlesLinksInteraction(tsParticles);
 
-    loadCircleShape(tsParticles);
-    loadImageShape(tsParticles);
-    loadLineShape(tsParticles);
-    loadPolygonShape(tsParticles);
-    loadSquareShape(tsParticles);
-    loadStarShape(tsParticles);
-    loadTextShape(tsParticles);
+    await loadCircleShape(tsParticles);
+    await loadImageShape(tsParticles);
+    await loadLineShape(tsParticles);
+    await loadPolygonShape(tsParticles);
+    await loadSquareShape(tsParticles);
+    await loadStarShape(tsParticles);
+    await loadTextShape(tsParticles);
 
-    loadLifeUpdater(tsParticles);
-    loadOpacityUpdater(tsParticles);
-    loadSizeUpdater(tsParticles);
-    loadAngleUpdater(tsParticles);
-    loadColorUpdater(tsParticles);
-    loadStrokeColorUpdater(tsParticles);
-    loadOutModesUpdater(tsParticles);
+    await loadLifeUpdater(tsParticles);
+    await loadOpacityUpdater(tsParticles);
+    await loadSizeUpdater(tsParticles);
+    await loadAngleUpdater(tsParticles);
+    await loadColorUpdater(tsParticles);
+    await loadStrokeColorUpdater(tsParticles);
+    await loadOutModesUpdater(tsParticles);
 }