+++ shapes/bubble/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.22.0](https://github.com/matteobruni/tsparticles/compare/tsparticles-shape-bubble@1.21.0...tsparticles-shape-bubble@1.22.0) (2021-10-28)
+
+**Note:** Version bump only for package tsparticles-shape-bubble
+
+
+
+
+
 # [1.21.0](https://github.com/matteobruni/tsparticles/compare/tsparticles-shape-bubble@1.20.4...tsparticles-shape-bubble@1.21.0) (2021-10-14)
 
 **Note:** Version bump only for package tsparticles-shape-bubble